Move `Preview` class to src/features/previewPanel.ts
