+++ src/features/previewPanel.ts
